+++ calcipy/tasks/types.py
@@ -3,11 +3,10 @@
 from beartype import beartype
 from corallium.file_helpers import read_package_name
 from invoke.context import Context
-from invoke.exceptions import UnexpectedExit
 
 from ..cli import task
 from ..invoke_helpers import run
-from .executable_utils import python_dir
+from .executable_utils import PYRIGHT_MESSAGE, check_installed, python_dir
 
 
 @beartype
@@ -20,13 +19,8 @@ def _inner_task(ctx: Context, *, command: str, cli_args: str = '') -> None:
 @task()
 def pyright(ctx: Context) -> None:
     """Run pyright."""
-    try:
-        _inner_task(ctx, command='pyright')
-    except UnexpectedExit as exc:
-        raise RuntimeError(
-            "pyright must be installed separately, such as 'brew install pyright' on Mac."
-            ' See the online documentation for your system: https://microsoft.github.io/pyright/#/installation',
-        ) from exc
+    check_installed(ctx, executable='pyright', message=PYRIGHT_MESSAGE)
+    _inner_task(ctx, command='pyright')
 
 
 @task()